refactoring of events processing
